+++ flag/service/legacyorganizations/legacyorganizations.go
@@ -1,0 +1,7 @@
+package legacyorganizations
+
+// LegacyOrganizations holds configuration specific to the oldschool
+// companyd organization service.
+type LegacyOrganizations struct {
+	Address string
+}